requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
psycopg[binary]==3.2.3
gunicorn==23.0.0

pytest==8.3.4
pytest-django==4.9.0

ruff==0.8.4
black==24.10.0
pre-commit==4.0.1



